more refactors, moved all the field processor out, will be updating by adding a 'getFieldsForFloat' method to it for the FloatFactory
