+++ src/main/java/group1/mips_simulator/components/Computer.java
@@ -4,7 +4,7 @@ package group1.mips_simulator.components;
 import group1.mips_simulator.components.cpuParts.CPU;
 import group1.mips_simulator.components.cpuParts.Register;
 import group1.mips_simulator.components.cpuParts.RomLoader;
-import group1.mips_simulator.components.dataParts.Field;
+import group1.mips_simulator.components.dataParts.FieldProcessors.Field;
 import group1.mips_simulator.components.dataParts.instructionParts.Instruction;
 import group1.mips_simulator.components.dataParts.instructionParts.InstructionFactory;
 import group1.mips_simulator.components.dataParts.instructionParts.RXIA_Instruction;
