+++ src/main/java/group1/mips_simulator/components/dataParts/FieldProcessors/FieldProcessor.java
@@ -1,7 +1,6 @@
-package group1.mips_simulator.components.dataParts.instructionParts.FieldProcessors;
+package group1.mips_simulator.components.dataParts.FieldProcessors;
 
 import group1.mips_simulator.Utility;
-import group1.mips_simulator.components.dataParts.Field;
 import group1.mips_simulator.components.dataParts.instructionParts.OpCode;
 
 import java.util.Vector;
