+++ src/main/java/group1/mips_simulator/components/dataParts/floatParts/Float.java
@@ -1,9 +1,12 @@
 package group1.mips_simulator.components.dataParts.floatParts;
 import group1.mips_simulator.components.dataParts.Value;
+import group1.mips_simulator.components.dataParts.FieldProcessors.Field;
 
 
 public class Float {
     protected Value value = new Value(0);
+    public Field[] fields = new Field[3];
+
     //in general, float data has the following fields, each with a purpose in calculating the float
     boolean signBit; //MSB (most significant bit, always indicates sign. Floats dont have an unsigned form)
     short bias; // the bits in the second field are converted to a unsigned int, then subtracted from the exponent bias
