+++ src/main/java/group1/mips_simulator/components/dataParts/instructionParts/RXIA_Instruction.java
@@ -1,6 +1,6 @@
 package group1.mips_simulator.components.dataParts.instructionParts;
 
-import group1.mips_simulator.components.dataParts.Field;
+import group1.mips_simulator.components.dataParts.FieldProcessors.Field;
 
 import java.util.Vector;
